requirements.txt
streamlit
plotly.express
pandas
os
warnings
numpy
seaborn
nltk
matplotlib.pyplot
re
time
pickle
math
tensorflow
json
requests
io
sys